Rename "cloud" package name part of local example to "local"
